+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added `IGMarkets::Errors::MarketOrdersNotSupported` and `IGMarkets::Errors::TooManyMarketsError`
 - Fixed the `ig_markets stream` command not finding the `curses` gem
+- The `curses` gem is now installed automatically on non-Windows platforms
 - Fixed the exception messages when trying to set an invalid or disallowed value on an attribute
 
 ### 0.25 — December 7, 2016